Change file from js to ts, add private variables
